+++ lib/src/livekit.dart
@@ -4,7 +4,7 @@ import 'room.dart';
 /// Main entry point to connect to a room.
 /// {@category Room}
 class LiveKitClient {
-  static const version = '0.5.3';
+  static const version = '0.5.4';
 
   /// Connects to a LiveKit room
   static Future<Room> connect(