+++ lib/src/publication/local_track_publication.dart
@@ -3,12 +3,15 @@ import '../extensions.dart';
 import '../internal/events.dart';
 import '../participant/local_participant.dart';
 import '../proto/livekit_models.pb.dart' as lk_models;
-import 'local_track.dart';
-import 'track.dart';
+import '../track/local.dart';
+import '../track/track.dart';
 import 'track_publication.dart';
 
-class LocalTrackPublication extends TrackPublication {
+class LocalTrackPublication<T extends LocalTrack> extends TrackPublication {
   final LocalParticipant _participant;
+
+  @override
+  covariant T? track;
 
   LocalTrackPublication(
     lk_models.TrackInfo info,