+++ lib/src/track/options.dart
@@ -43,12 +43,19 @@ class CameraTrackOptions extends LocalVideoTrackOptions {
       );
 }
 
-class ScreenTrackOptions extends LocalVideoTrackOptions {
-  const ScreenTrackOptions();
+class ScreenShareTrackOptions extends LocalVideoTrackOptions {
+  const ScreenShareTrackOptions();
+}
+
+/// Base class for track options.
+abstract class LocalTrackOptions {
+  const LocalTrackOptions();
+  // All subclasses must be able to report constraints
+  Map<String, dynamic> toMediaConstraintsMap();
 }
 
 /// Options when creating a LocalVideoTrack.
-abstract class LocalVideoTrackOptions {
+abstract class LocalVideoTrackOptions extends LocalTrackOptions {
   // final LocalVideoTrackType type;
   final VideoParameters params;
 
@@ -56,6 +63,7 @@ abstract class LocalVideoTrackOptions {
     this.params = VideoParameters.presetQHD169,
   });
 
+  @override
   Map<String, dynamic> toMediaConstraintsMap() => <String, dynamic>{
         'mandatory': params.toMediaConstraintsMap(),
       };
@@ -228,13 +236,41 @@ class VideoParameters {
   // https://developer.mozilla.org/en-US/docs/Web/API/MediaDevices/getUserMedia
   //
   Map<String, dynamic> toMediaConstraintsMap() => <String, dynamic>{
-        'maxWidth': width,
-        'maxHeight': height,
+        'minWidth': width,
+        'minHeight': height,
         'maxFrameRate': encoding.maxFramerate,
       };
 }
 
 /// Options when creating an LocalAudioTrack. Placeholder for now.
-class LocalAudioTrackOptions {
-  const LocalAudioTrackOptions();
+class LocalAudioTrackOptions extends LocalTrackOptions {
+  final bool noiseSuppression;
+  final bool echoCancellation;
+  final bool autoGainControl;
+  final bool highPassFilter;
+  final bool typingNoiseDetection;
+
+  const LocalAudioTrackOptions({
+    this.noiseSuppression = true,
+    this.echoCancellation = true,
+    this.autoGainControl = true,
+    this.highPassFilter = false,
+    this.typingNoiseDetection = true,
+  });
+
+  @override
+  Map<String, dynamic> toMediaConstraintsMap() => <String, dynamic>{
+        'optional': <Map<String, dynamic>>[
+          <String, dynamic>{'echoCancellation': echoCancellation},
+          <String, dynamic>{'googDAEchoCancellation': echoCancellation},
+          <String, dynamic>{'googEchoCancellation': echoCancellation},
+          <String, dynamic>{'googEchoCancellation2': echoCancellation},
+          <String, dynamic>{'noiseSuppression': noiseSuppression},
+          <String, dynamic>{'googNoiseSuppression': noiseSuppression},
+          <String, dynamic>{'googNoiseSuppression2': noiseSuppression},
+          <String, dynamic>{'googAutoGainControl': autoGainControl},
+          <String, dynamic>{'googHighpassFilter': highPassFilter},
+          <String, dynamic>{'googTypingNoiseDetection': typingNoiseDetection},
+        ],
+      };
 }